% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{regexMatch}
\alias{regexMatch}
\title{C++ regex_match wrapper}
\usage{
regexMatch(s, m)
}
\value{
string
}
\description{
C++ regex_match wrapper
}
\examples{
regexMatch("1:2:last_name", "(.*):(.*):last_name")
}

