% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrocksdb-package.R
\docType{package}
\name{rrocksdb-package}
\alias{rrocksdb-package}
\alias{rrocksdb}
\title{rrocksdb}
\description{
rrocksdb
}
\author{
Scott Chamberlain \email{sckott@protonmail.com}
}
\keyword{package}
